+++ merchant.php
@@ -156,7 +156,7 @@ class merchant extends ecjia_merchant
         $price_ladder = !empty($_POST['price_ladder']) ? $_POST['price_ladder'] : '';
         $restrict_amount = !empty($_POST['restrict_amount']) ? $_POST['restrict_amount'] : '';
         $gift_integral = !empty($_POST['gift_integral']) ? $_POST['gift_integral'] : 0;
-        $deposit = (!empty($_POST['deposit']) && intval($_POST['deposit']) > 0) ? intval($_POST['deposit']) : 0;
+        $deposit = (!empty($_POST['deposit']) && floatval($_POST['deposit']) > 0) ? floatval($_POST['deposit']) : 0;
 
         $price_ladder = array();
         $count = count($_POST['ladder_amount']);
@@ -515,7 +515,7 @@ class merchant extends ecjia_merchant
             $price_ladder = !empty($_POST['price_ladder']) ? $_POST['price_ladder'] : '';
             $restrict_amount = !empty($_POST['restrict_amount']) ? $_POST['restrict_amount'] : '';
             $gift_integral = !empty($_POST['gift_integral']) ? $_POST['gift_integral'] : 0;
-            $deposit = (!empty($_POST['deposit']) && intval($_POST['deposit']) > 0) ? intval($_POST['deposit']) : 0;
+            $deposit = (!empty($_POST['deposit']) && floatval($_POST['deposit']) > 0) ? floatval($_POST['deposit']) : 0;
 
             $price_ladder = array();
             $count = count($_POST['ladder_amount']);
